+++ pages_template/uni-id/login/forget/forget.vue
@@ -10,7 +10,7 @@
 				<view class="form-item form-border">
 					<!-- 文本框 -->
 					<u-input
-						class="main-input"
+						class="form-input"
 						v-model="form1.mobile"
 						type="text"
 						:maxlength="11"
@@ -22,7 +22,7 @@
 				<view class="form-item form-border">
 					<!-- 文本框 -->
 					<u-input
-						class="main-input"
+						class="form-input"
 						v-model="form1.password"
 						type="password"
 						placeholder="请输入新密码"
@@ -32,7 +32,7 @@
 				<view class="form-item form-border">
 					<!-- 文本框 -->
 					<u-input
-						class="main-input"
+						class="form-input"
 						v-model="form1.password2"
 						type="password"
 						placeholder="请再次输入新密码"
@@ -44,7 +44,7 @@
 				<view class="form-item form-border">
 					<!-- 文本框 -->
 					<u-input
-						class="main-input"
+						class="form-input"
 						v-model="form1.code"
 						type="number"
 						:maxlength="6"
@@ -175,7 +175,7 @@
 					success:function(data){
 						that.isRotate = false;
 						vk.alert("密码重置成功,新密码为:"+password,function(){
-							// 进入首页
+							// 进入登录页
 							vk.redirectTo("../index/index");
 						});
 					},