+++ pages_template/uni-id/login/index/index.vue
@@ -9,38 +9,25 @@
 
 				<view class="form-item form-border">
 					<!-- 文本框 -->
-					<u-input
-						class="main-input"
-						v-model="form1.username"
-						type="text"
-						:maxlength="50"
-						placeholder="用户名/手机号"
-						placeholder-style="'color':'#8e8e8e'"
-					/>
+					<u-input class="form-input" v-model="form1.username" type="text" :maxlength="50" placeholder="用户名/手机号" placeholder-style="'color':'#8e8e8e'" />
 				</view>
 				<view class="form-item form-border">
 					<!-- 文本框 -->
-					<u-input
-						class="main-input"
-						v-model="form1.password"
-						type="password"
-						placeholder="密码"
-						placeholder-style="'color':'#8e8e8e'"
-					/>
+					<u-input class="form-input" v-model="form1.password" type="password" placeholder="密码" placeholder-style="'color':'#8e8e8e'" />
 				</view>
 			</view>
 			<view class="login-btn">
 				<u-button shape="circle" @click="login" :plain="false" :hair-line="false" type="success">登 录</u-button>
 			</view>
 			<!-- 其他登录 -->
-			<view class="other_login cuIcon">
+			<view class="login-icon-view">
 				<!-- #ifdef MP-WEIXIN -->
-				<view class="login_icon">
+				<view class="login-icon-item">
 					<u-icon name="weixin-fill" @click="login_weixin" size="80" color="#19be6b"></u-icon>
 				</view>
 				<!-- #endif -->
 				<!-- #ifdef APP-PLUS -->
-				<view class="login_icon">
+				<view class="login-icon-item">
 					<u-icon name="weixin-fill" @click="login_weixin" size="80" color="#19be6b"></u-icon>
 				</view>
 				<!-- #endif -->
@@ -184,7 +171,7 @@
 					vk.navigateBack();
 				}else{
 					// 进入首页
-					vk.reLaunch("../../index/index");
+					vk.navigateToHome();
 				}
 			},
 			// 第三方登录 - 微信